apps/backend/sentinel/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-backend"
version = "0.1.0"
description = "Sentinel agent backend (FastAPI)"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "sqlalchemy>=2.0.35,<3.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pgvector>=0.3.6,<1.0.0",
  "PyJWT>=2.9.0,<3.0.0",
  "httpx>=0.27.0,<1.0.0",
  "croniter>=1.4.0,<2.0.0",
  "pyyaml>=6.0.0,<7.0.0",
  "tiktoken>=0.7.0,<1.0.0",
  "playwright>=1.52.0,<2.0.0",
  "websockets>=12.0,<13.0",
  "python-telegram-bot>=21.0,<22.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "black>=25.1.0,<26.0.0",
]

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
